添加需求模块接口

+++ nyc-framework/nyc-common/src/main/java/com/git/bds/nyc/page/PageParam.java
@@ -13,18 +13,18 @@ import java.io.Serializable;
 @Data
 public class PageParam implements Serializable {
 
-    private static final Integer PAGE_NO = 1;
-    private static final Integer PAGE_SIZE = 10;
+    private static final long PAGE_NO = 1;
+    private static final long PAGE_SIZE = 10;
 
     @ApiModelProperty(value = "页码，从 1 开始", required = true,example = "1")
     @NotNull(message = "页码不能为空")
     @Min(value = 1, message = "页码最小值为 1")
-    private Integer pageNo = PAGE_NO;
+    private Long pageNo = PAGE_NO;
 
     @ApiModelProperty(value = "每页条数，最大值为 100", required = true, example = "10")
     @NotNull(message = "每页条数不能为空")
     @Min(value = 1, message = "每页条数最小值为 1")
     @Max(value = 100, message = "每页条数最大值为 100")
-    private Integer pageSize = PAGE_SIZE;
+    private Long pageSize = PAGE_SIZE;
 
 }
